[REM-103] refactor: update configuration and decorator imports
refactor:
- config.py: removed redundant directory removal and ensured directory creation for ARCHIVES_PATH
- decorators/__init__.py: added current_user to exports
- routes/__init__.py: updated import path for current_user and login_required
