+++ web/decorators/__init__.py
@@ -1,6 +1,6 @@
 """Module for importing and managing decorators in the application."""
 
 from .checks import check_privilegies
-from .login_wrap import login_required
+from .login_wrap import current_user, login_required
 
-__all__ = ["check_privilegies", "login_required"]
+__all__ = ["check_privilegies", "login_required", "current_user"]
